Removed 2 tags, added length limits to ship-to's
Removing (in header): this tag is not necessary for Shopify, so removing it completely from the header. <RequestedDeliveryDate></RequestedDeliveryDate>

Updating (in header): 
Change warehouse code from abbreviated Whs to spell out Warehouse like this: <Whs>VRN</Whs> to <Warehouse>VRN</Warehouse>. This does not apply to the details section, where we DO use the abbreviated <Whs> warehouse tag.